A unit test for vanilla mlps
Summary:
What: This is a simple test to check if vanilla mlps are able to learn a custom function.

Why: The need for a unit test to check instantiation of mlps, and their learnability was felt.

How: the training data is generated with x sampled from a multivariate normal distribution and y as the corresponding pdf values. A vanilla mlp is trained to learn this mapping. The training loss (mse error) should go down with training steps.

Differential Revision: D47030273

fbshipit-source-id: d5abfb8142b9153d6e3f61160b749365df1a6a6a