use get_batch_action_value method in q networks to run all temporal difference based algorithms
Summary:
What is this diff: This diff attempts to get rid of using the forward method in training rl algorithms. Changes are made to the base deep_td_learning class and all child classes. Also, 'get_batch_action_value' method has been moved out of the VanillaValueNetwork class (since this is just a plain mlp) and put in value network classes (VanillaStateActionValueNetwork, DuellingStateActionValueNetwork)

Why is this diff required: Currently, we sometimes use the forward pass and some times use the get_batch_action_value method; this was introduced due to algorithmic computations required for duelling dqn (we need q values of all available actions to compute the mean of advantage and subtract it). I think a clean way is to restrict the forward method to the neural network architecture and instead use get_batch_action_value (a method in q networks) to interface with all algorithms.

How are the changes made: To deal with the unique computations required by duelling dqn, we make 'curr_available_action_batch' as an optional input to the 'get_batch_action_value' method. For all other algorithms (dqn, double dqn, sarsa, pg) this is set to none and the 'get_batch_action_value' method is equivalent to a forward pass. For dueling, whenever we pass a tensor to 'curr_available_action_batch', computations are done to get the mean value

Differential Revision: D46745190

fbshipit-source-id: 7d66e229d1e1c19d09e46e48379fe8607f04555d