Move Cogwheel tests to integration tests, and get Cogwheel tests to dynamically copy those.
Summary:
The main integration tests in Pearl have so far been running as Cogwheel tests. The use of Cogwheel tests is due to the fact that they can be used on Sandcastle, upon submission of Diffs, without timing out (regular tests would time out).

Like all the rest of Pearl code, Cogwheel tests were being exported the open source distribution, which should not happen because Cogwheel is Meta-specific. We can avoid that by moving Cogwheel tests to a `meta_only` directory, which by convention is not exported to the open source distribution.

However, these tests are important and it is desirable that they be available to open source users and developers. Since the open source distribution cannot rely on Cogwheel, they must be available as regular tests (using the `unittest` library). More specifically, they must be _integration_ tests since they test multiple parts of the library at once. This diff moves these tests to a new `test/integration` directory, where they are defined with `unittest`. This ensures their presence in the open source distribution as integration tests.

However, we still want to run them as Cogwheel tests _inside_ Meta. To achieve that, we could keep the `cogwheel` directory as it was, but that would imply duplicating the integration tests defined in `integration`. This is undesirable since adding and modifying them would require people to keeping them in sync.

To avoid this duplication, this diff rewrites the Cogwheel tests code to _dynamically_ reuse the `integration` test definitions, wrapping them up as Cogwheel tests. This is done by inspection code that actually iterates over the methods of the integration test class and copies them into a CogwheelTest class, one by one. This way, we only need to define long-running tests in `integration`, and they will be automatically run as Cogwheel tests on Sandcastle.

The diff also defines a Buck target `test_pearl_integration` for the `unittest` integration tests in case developers want to run them locally with the `--timeout` flag to avoid timing out. This target is tagged as `disabled` to avoid its execution on Sandcastle.

Finally, the diff also renames directory `unittest`, which contained the unit tests, to `unit`, since `integration` is also defined via `unittest`. Having a pair of directories `unit`/`integration` makes the distinction between unit and integration tests clear, and does not make it look like only one of them is using the `unittest` library.

Differential Revision:
D47614770

Privacy Context Container: L1202097

fbshipit-source-id: 714e961259e96188c3a2750edc4f8df618c92d0c