fix for ppo optimizer
Summary:
Why?
This is caused by a diff racing between ppo and sac
sac renamed the optimizer in PG, but due to how codes are written, this didnt cause any code conflict.

it made ppo performance much worse and flaky on cogwheel test.

How?
added a helper function in PG to set optimizer to avoid this from happening again in the future.

Reviewed By: BillMatrix

Differential Revision: D47611547

fbshipit-source-id: 67b4489a60a743ad43f9e8cf6938584db4d37c99