Redo of conservative q learning (algorithm part) - copy of D47386414
Summary:
Redo of the conservative q learning algorithm part as the previous commit D47386414 got messed up with file moves.
This is only the algorithmic part - offline data collection and training pipeline will be added in a separate commit but will follow the strategy in this notebook: N3981388. Also, see here the outperformance of cql vs vanilla dqn on offline data.

Per previous comments, compute_cql_loss is added as a separate utility function so that it can be reused by other offpolicy algorithms. Minimal changes are made to the base td learning class for clean expositon.

Reviewed By: BerenLuthien

Differential Revision: D47702225

fbshipit-source-id: eca34a0dd6efb1da79c8a610e2cd38ae5960b415