offline testing pipeline for cql
Summary:
offline testing for cql, ready for review.

Summary

Notes
Even though it passed local integration test with a target of 100, I changed the target to 50 to avoid any flakiness. Flakiness is expected for reinforcement learning algorithms trained with offline data. I am also training with a smaller data size (50k transition tuples). A target of 50 should be easy for cql with this data set

Reviewed By: rodrigodesalvobraz

Differential Revision: D48169862

fbshipit-source-id: 64651223f08b0936e4ff08dbb422dd8321cc767f