Rename StateActionValueNetwork to QValueNetwork
Summary:
This diffs renames `StateActionValueNetwork` to the equivalent, shorter, and more standard `QValueNetwork`.

It also renames method `get_batch_action_value` to the more consistent (with the class name) name of `get_q_values`. Note that `NpletsCritic` already uses this name, so this is a step towards making `NpletsCritics` a subclass of `QValueNetwork`.

Differential Revision: D48282358

fbshipit-source-id: ca0fee2032cc6d703a2947ea2403a73cbff1c335