Introduce StateActionValueNetwork type
Summary:
Even though we had more than one type of state-action value network, there was no `StateActionValueNetwork` type. This was needed for proper type hinting in the `NpletsCritic` class which contains a list of state-action value networks.

This diff introduces this type and the methods it requires, and uses it for correct typing in NpletCritic. In the process, it was necessary to introduce the also so-far-inexistent `ValueNetwork` type, which must be an ancestor of `StateActionValueNetwork` as well of `VanillaValueNetwork`.

Differential Revision: D48258917

fbshipit-source-id: 5054a3872e5c72094bcd3243ceab45f30dd988d4