implicit q learning
Summary:
IQL implementation. It uses QValueNetworks with layer normalization and random ensemble distillation (instead of clipped double q learning) as modifications to the original implementation. See this paper: https://arxiv.org/pdf/2302.02948.pdf

Notebook for testing: N4173357

Changes also made to add twin critic class implementation.

Why the need for Twin Critic as compared to the one we have? The current implementation of Twin Critic wraps things too deep making it unclear; the user might be forced to read and understand it in order to understand different algorithm implementations.

1) It automatically creates target networks; so just looking at its used in algorithm implementation, it is not clear whether target networks are being used or not.

2) The optimizers are added to a list and the 'optimize' function performs the optimization step sequentially. A better idea to do this is to use torch.optim and put different parameters in a dictionary. Since typically the optimization objective is sum of individual losses, one optimization step should suffice. Moreover, the use of get_q_value_estimate_fn seems unintuitive at first pass: we probably do not expect something different than a forward pass though the q networks to get q values. Then, why should this be an input to the optimize function?

3) The 'get_q_values' function is not super clear. Because the class has both the base and target networks defined together, we need to specify from which set of networks we need to get q values from.

4) The 'update_target_networks' function can be used more generally. Hence, I provided a function separately. For examples, this might be useful when having an ensemble. I will try to see if the 'update_target_networks' (updating a collection of networks) I wrote can replace 'update_target_network' (updating one network) more generally. I agree with you, ideally, we should have only one update function.

5) Why two different functions, 'get_q_values' and 'forward' when both are doing forward pass from the critics? Also, I think we should not use the 'forward' function in ValueNetworks or QValueNetworks class and instead always use the 'get_q_values' function.

The way I wrote Twin Critic class: Twin Critic, as the name suggests, is simply a collection of two critics; has a function which updates the critics towards a target and another function which gets values from both.

It makes things super clear when using in algorithm implementation; please see IQL as an example. It does not automatically create targets; one has to define the target networks as another instance, making it clear that the algorithm uses target networks. It forces algorithm implementation to make it clear whether a sample is taken from base networks or target networks (see lines 150-151 or 201-202). It forces one to show how the target networks are updated; whether at every step or every few steps. The use of function 'update_target_networks' in learn_batch clarifies it.

Notes:
I could have modified the NpletsCritic class but since ddpg and td3 algorithms already use this, I thought to propose my changes first applied to IQL. If this is ok, I can rebase ddpg and td3 on the way I wrote Twin Critic.

I do think a general class for ensemble of networks is useful -  which can be leveraged by ensemble dqn and ensemble epistemic networks implementation. I can generalize the Twin Critic class in a way that it can be reused for exploration algorithms as well.

Reviewed By: rodrigodesalvobraz

Differential Revision: D48474395

fbshipit-source-id: 0969ba245150c02c731a524b517df4c06a791e42