support keyed optimizer in pearl
Summary:
Why:
APS has requirement on type of optimizer to be KeyedOptimizer.
Therefore adding an input config to change this in pearl.

What:
- add use_keyed_optimizer bool to indicate optimizer type. (NOTE: PEARL has a discussion to switch to config file, along with that effort, we need to define optimizer config as well, then we could get rid of this bool flag)

jarodhou has confirmed that changes in PEARL on this stack is good enough to unblock him continue on APS side.

BillMatrix confirmed that adding this keyed optimizer in PEARL is ok.
 {F1091215840}

Reviewed By: BillMatrix

Differential Revision: D49277012

fbshipit-source-id: 956f99d27f688af97c2212ad37e6a7e2b04999d0