Fix GymEnvironment to avoid passing tensor-represented actions to wrapped Gym environment
Summary:
Our Gym environments wrapper was passing actions to the `step` method even if the action was a PyTorch tensor.

This worked in earlier versions of NumPy but not in later versions, causing an exception.

To avoid that, this Diff changes the  wrapper to convert tensors to ndarrays.

This change reveals a former problem in Pearl because now the rewards being output from Gym environments are floats (they were tensors before), and when pushed to tensor-based replay buffers are converted to a tensor which had default Double (float64). This raises an exception when running `loss.backward()` because some tensors are Float and some are Double. To fix this, tensor-based replay buffer now ensures the reward back is Float.

Reviewed By: BillMatrix

Differential Revision: D49469395

fbshipit-source-id: b8ccfa31a7d0d0ed2973e503a279a2fb2172b820