Fix ddpg and continous SAC act methods
Summary:
Our Gym environments wrapper was passing actions to the `step` method even if the action was a PyTorch tensor.

This worked in earlier versions of NumPy but not in later versions, causing an exception.

To avoid that, we will change the  wrapper to convert tensors to ndarrays.

However, this revealed a previous problem in some Pearl methods (ddpg and continuous SAC), which computed *batches* of actions even when receiving a single, non-batch state. This (single-action) batch of actions was being passed all the way to Gym environments. When these batches were tensors, this somehow worked because the extra batch dimension was ignored, but when it is converted to ndarray the extra dimension is not ignored and the resulting rewards are also output by the environment as a batch, causing issues down the line because Pearl assumes these rewards are not batched.

This diff fixes ddpg and continous SAC methods to _not_ output a batch of actions when receiving a non-batch state.

Reviewed By: BillMatrix

Differential Revision: D49469394

fbshipit-source-id: 1989cc499f67d918717c62d9b7e64e49010d44db