flake8 config
Summary:
Add flake8 config to the main pearl directory with max-line-length = 88, which is the same as the black formatter. The default fbsource .flake8 config turns off the max-line-length warning, but this will eventually be raised in open source. Therefore, we should have our own flake8 config here.

I used the settings from `fbcode/ax/.flake8`, which are reasonable and ensure good code quality.

Reviewed By: BillMatrix

Differential Revision: D50177384

fbshipit-source-id: 8d662a2a7c493905530b51847e22f25c27bb1326