Register buffers in quantile value network
Summary:
`QuantileQValueNetwork` was using tensor attributes not registered as buffers, causing them not to be moved to a device when using method `to()`.

This diff registers them as buffers. It also exposes `quantile_midpoints` at the `DistributionalQValueNetwork` because `QuantileRegressionDeepTDLearning` accesses that property from its `_Q` network, which is only required to be `DistributionalQValueNetwork`.

Reviewed By: jb3618columbia

Differential Revision: D50779460

fbshipit-source-id: 8b297238297fc464c073bfb022aefef45a573803