Remove PyTorch device determination from everywhere but PearlAgent
Summary:
Previously, most of Pearl depended on global function `get_pearl_device` to look up which device to move tensors to before performance-intensive operations.

However, it may be the case that client code (in our case, Meta's internal APS framework) wishes to determine that device itself, move a `PolicyLearner` to that device using `nn.Module.to(device)`, and provide data in the form of tensors already in the chosen device. These requirements excludes `PearlAgent` and replay buffers.

We explicitly codify these requirements in the following way:
A. `PolicyLearner` is an `torch.nn.Module` and therefore can be moved to a new device with method `to(device)`.
B. Any inputs to a policy learner must already be in the same device as the models that will process them.

This diff makes changes to satisfy these requirements. It does so by eliminating calls to method `to` for input tensors inside policy learners and their components, since all input must already be in the correct device. Temporary tensors built inside policy learners are moved to the same device as the other input.

Note that not all policy learners require tensors, so a new attribute `require_tensors` (default `True`) is added to `PolicyLeaner` to inform client code whether inputs are supposed to be converted to tensors and moved to a device.

Reviewed By: alexnikulkov

Differential Revision: D50827882

fbshipit-source-id: d5ffcb45ad68ed30abe4f5bc3627e84fee6fa0a8