Action Representation Module and One Hot Representation
Summary:
There remains a key blocker on how to consistently transform raw actions into representations that can be leveraged by neural networks.

In this diff, we build the foundation for action representation modules that takes in raw actions and outputs representations. We also introduce onehot representations here for discrete actions that only provides a single index for action.

Reviewed By: danielrjiang

Differential Revision: D51131215

fbshipit-source-id: ed464cf1789110c2d07ab5516bbc51801d804c3d