`GymEnvironment` + Pearl action spaces
Summary:
Main changes are in `gym_environment.py` and `environment.py` (all others are pyre ignores that will be fixed later).

In `gym_environment.py`:
* Added a mapping that maps Gym action space names to Pearl action classes (`GYM_TO_PEARL_ACTION_SPACE`). This is used when returning the action space. If the original Gym env has a Discrete space, then we return a Pearl `DiscreteActionSpace`. Similar for `Box`.
* Similarly, added a mapping `PEARL_TO_GYM_ACTION` that maps Pearl actions (Tensors) to the format expected by Gym. I.e., Gym discrete action spaces expect an integer, so we convert a single-element Tensor to an `int`.
* I removed the `seed` function because this is not supported in the new Gymnasium (it's supposed to be passed into the `reset` only).

NOTE: How critical is it that we continue to support old Gym or Gymnasium 21? Wondering if we should just take out some of the dual functionality.

NOTE: The new pyre-ignores are due to adding a typehint to `GymEnvironment.observation_space`.

Reviewed By: rodrigodesalvobraz

Differential Revision: D51369688

fbshipit-source-id: c9d22061acec9537b293f10a17913e41ccc83f85