fixing entropy autotuning in sac
Summary:
Entropy auto tuning does not work for soft actor critic algorithms because the learnable parameter '_log_entropy_' was registered as a buffer preventing the optimizer from updating it. This has been modified and the learnable parameter is registered as a parameter instead allowing the optimizer to update it.

learn_batch method has been modified to included the entropy autotuning code, instead of writing it in actor_learn_batch. I somehow missed this when landing my previous diff with modifications to the actor critic base class. BillMatrix

Reviewed By: BillMatrix

Differential Revision: D51518898

fbshipit-source-id: 7e45d16e813c2bca335af7ac9f47c881ce41cba6