Don't skip empty batches in disjoint CB training
Summary:
Due to asymmetrical distributed data loading some of the batches might not have all arms present. This breaks distributed training because some of the all_reduce calls are unmatched. This diff removes skipping of empty batches, so that for each batch and each arm we make an all_reduce call (0 values are used for actions which aren't present in the batch).

Changes in this diff:
1. Remove skipping of empty batches in disjoint CB. When the batch is empty, replace it with a null batch. Null batch has a single element with 0 weight and dummy state/action/reward values.

Differential Revision: D53028375

fbshipit-source-id: ea282267a01d2e581b4779d8f49383e0ac7da03d