Correcting small bugs in the offline learning environment
Summary:
I noticed some small bugs in the offline rl pipeline. This diff corrects that.

Primarily, we need an attribute called `max_number_actions` when creating a transition tuple.  See for example the `push` function in `FIFOOffPolicyReplayBuffer`. This is required for using Pearl with discrete action spaces and done to enable Pearl to work with dynamic action spaces.

However, I think the current implementation for offline rl pipeline is not appropriate. Specifically,

- The 'create_offline_data' function creates transition tuples with an attribute called `max_number_actions`. However, offline data already collected by users may not contain this attribute. We want to be consistent. Even when creating our own offline data for testing, we should not add this attribute.

- The `get_offline_data_in_buffer` function in 'offline_learning_and_evaluation.py' file has a strange way of computing the `max_number_actions`. First, the transition in the offline date may not contain a field called `action_space` (note, the transition tuples we collect when learning offline do not have the `action_space` attribute. Second, for continuous action spaces, we do not need `max_number_actions`.

I also added more documentation along the way.

Reviewed By: rodrigodesalvobraz

Differential Revision: D54273885

fbshipit-source-id: c80f0863094f78c10eaa85c5284332bd88550432