switch squeeze() to squeeze(-1) in Pearl CB model output
Summary: We had problems with the `squeeze()` unexpectedly squeezing out the batch dimension when batch_size=1. This squeeze was originally intended to squeeze the arm dimension when we have a joint model or are making the predictions for a single arm. Using `squeeze(-1)` will make sure that we only squeeze the arm (last) dimension and keep the batch dimension intact

Reviewed By: danielrjiang

Differential Revision: D55025833

fbshipit-source-id: de713b3a946e7d26d352c31908120a8b223af537