Make Tabular Q-learning deterministic modulo exploration
Summary:
Tabular Q-learning broke ties between actions randomly. This non-determinism is undesirable because it is present even when exploration is turned off, making it harder to work with it in some contexts (for example, checking the equality between two such agents by checking the equality of their outputs to a sample of observations).

So now it selects the first best action.

Note that this hurts the performance of the policy learner, because the random tie-breaker between actions of the same value worked as a useful exploration. However, exploration should be confined to the exploration module. If we want to have that kind of exploration, we must instead use an exploration module that does uses that tie-breaking (something similar to propensity exploration).

Reviewed By: yiwan-rl

Differential Revision: D54989173

fbshipit-source-id: caca0522b266517d4af627b8ba1e2eea396f794e