TS bug fix
Summary:
TS bug fix: numerical errors leads to deltaA to be non symmetric (order small <1e-8). When summing this it may result with a non-symmetric matrix A which leads to an error in the type check: when sampling from a Gaussian distribution the covariance matrix is required to be both PD and symmetric.

The offered solution is to explicitly symmetrize it, hence guarantee deltaA is symmetric.

Reviewed By: danielrjiang

Differential Revision: D56936024

fbshipit-source-id: 8219ca1e1a1577b0a241c0a2d7848ab02f427c2c