Store reply buffers in CPU only
Summary:
So far, every tensor added to a tensor-based replay buffer was moved to the active device, leading the entire replay buffer to be stored in it. For large replay buffers this will be too much for the GPU.

Now we store the replay buffer in CPU and move batches sampled from it to the active device.

Reviewed By: yiwan-rl

Differential Revision: D54625530

fbshipit-source-id: 9afe8e65c33077621a527b3a0cc061244854119f