Add option to force use of pinv in LinUCB (#89)
Summary:
Pull Request resolved: https://github.com/facebookresearch/Pearl/pull/89

Adding an options to force pseudo-inverse to be used by default in LinUCB. This is useful for cases where the features are strongly correlated, so the design matrix is ill-conditioned.
The regular inverse is faster, it should be used when matrix inversion is the bottleneck and the design matrix is well-conditioned.

Reviewed By: rodrigodesalvobraz

Differential Revision: D57462061

fbshipit-source-id: 4aaa18b1aa75d21e75561e058a711d27ad0f7d7e