add action representation to bootstrapped DQN
Summary:
The current implementation of bootstrapped DQN encounters a dimension mismatch error when integrated with an action representation module, such as the OneHotActionTensorRepresentationModule used in CartPole-v1. This issue arises because the deep exploration mechanism fails to incorporate the action representation module during processing. This update resolves this issue.

Additionally, this update modifies the default parameters in our benchmark settings for bootstrapped DQN. Specifically, the prior scale has been adjusted from 0 to 100, and the number of elements in the Q ensemble network has been reduced from 10 to 5. The absence of a prior network (0 prior scale) results in all ensemble elements ranking actions similarly after a few training steps, leading to no exploration and, therefore, no performance improvements. Reducing the number of ensembles accelerates the training process.

Reviewed By: rodrigodesalvobraz

Differential Revision: D60557188

fbshipit-source-id: 57ec2181703d973ec59af65751794b319ed3d8fd