Factor out replicated code in replay buffers
Summary:
All replay buffer implementations were replicating code for padding action vectors. This makes the code more confusing, and harder to maintain.

This diff solves this by moving that functionality to the super class `TensorBasedReplayBuffer` level. Now `push` is defined there, and delegates the computation and storage of the transition (the only part in which the implementations differ) to a new protected method `_store_transition`. Implementation define only that method now, based on their specific methods.

The exception to that is `HindsightExperienceReplayBuffer`, which overrides `push` with some very different functionality.

Reviewed By: yiwan-rl

Differential Revision: D61639099

fbshipit-source-id: 8303c06f8cadd8d25b9ff8e8d9570b3937e872f0