Make Generic Replay Buffer
Summary:
`OnPolicyReplayBuffer` is an ad hoc type of replay buffer for adding some extra fields to transitions.

In fact, it adds two distinct sets of fields to serve two distinct algorithms (REINFORCE and PPO). It would be better to have one type of replay buffer per algorithm instead, each one adding only the necessary fields.

This however would result in two very similar dataclasses where the only difference would be the number and names of added fields.

This diff generalizes the code in `OnPolicyReplayBuffer` to a function that generates a replay buffer class based on given transition and transition batch classes. In the next diff this utility is used to generate replay buffer classes for both REINFORCE and PPO, without code duplication.

Reviewed By: jb3618columbia

Differential Revision: D62413950

fbshipit-source-id: 5dd59193426099920c49d1fc8a46ea978b119b9e