Remove unnecessary DiscreteContextualBanditReplayBuffer class
Summary:
Pearl defined a replay buffer for contexual bandits that was rather basic, `DiscreteContextualBanditReplayBuffer`.

This was probably defined because regular replay buffer take many more fields than necessary for a contextual bandits.

However, with the recent refactoring of replay buffers, it is possible to use `BasicReplayBuffer` for discrete contextual bandits.

This diff slightly extends `TensorBasedReplayBuffer` to not require `curr_available_actions` (it is optional in transitions anyway), which allows us to use its subclass `BasicReplayBuffer` for contextual bandits.

It then also removes `DiscreteContextualBanditReplayBuffer` and replaces its usages by `BasicReplayBuffer`.

Reviewed By: Yonathae

Differential Revision: D62582430

fbshipit-source-id: e09546cdac1ba09c77994cd83b1a35aa0ec70766