change observations only from float64 to float32 before adding to the replay buffer
Summary: This is to prepare future diffs for supporting Atari games. In Atari games, the observations are uint8 type. Currently, pearl converts uint8 observations to float32 before adding them to the replay buffer. A huge amount of memory is used when the agent saves float32 images in the replay buffer. This diff reduces the memory usage by not making the conversion when the observations are added to the replay buffer, and makes this conversion when batchs are sampled from the replay buffer.

Reviewed By: rodrigodesalvobraz

Differential Revision: D65923483

fbshipit-source-id: b43825a31c4da2aa4e58cea5bb47025379b0903e