Make all Pearl components torch.nn.Modules
Summary:
This diff makes all Pearl components a subclass of `nn.Module`.

This works fine except property setters for modules do not work in classes which are modules themselves (https://stackoverflow.com/q/79356290/3358488), so we also need to adjust that. This consists of deleting setters, and obtaining the value of properties from `self._modules` rather than a protected attribute.

Reviewed By: Yonathae, yiwan-rl

Differential Revision: D67707011

fbshipit-source-id: 95096a8d01db7b12e5e1fdd631ec5042a3cb6fe2