Add one-hot encoding for LinearBandit
Summary:
LinearBandits attach the action index to the features used by LinearRegression, but this implies that the actual numerical value of the action ends up mattering (an action who index is twice that of another index will have twice the effect).

To avoid this spurious effect, we modify the code to use a one-hot encoding for the action.

A new parameter `one_hot_action_representation_CB` is introduced for the workflow, with a default value `True`.

Reviewed By: PavlosApo

Differential Revision: D75389940

fbshipit-source-id: e4dedf64aa30c603b7fa665021acee83f9116a62