Correct ContextualBanditBase.get_score signature in itself and subclasses
Summary:
Fixes inconsistent signatures of `get_scores` in `ContextualBanditBase` and its subclasses, which had a new action space parameter. The fix is to add that parameter to the super class as well.

Also renaming the parameter to `action_space_to_score` to stress that this is not necessarily the entire action space.

Differential Revision: D75363446

fbshipit-source-id: 017cfa57f96f5da151af0923db7e1483a59d03ba