Improve zero coefficient test for linear regression
Summary: This diff improves the zero-coefficient test for linear regression in the `test_linear_regression.py` file. The code changes include setting up ground truth based on a Linear Bandit application where the feature is the concatenation of state and one-hot action encoding. The zero-coefficient property is verified after each `learn_batch` operation, not just at the end of training. The diff also includes a list of zero-feature indices to be used in the test.

Reviewed By: Yonathae

Differential Revision: D76729858

fbshipit-source-id: 812be82a668d07fd078732b86e856b2dc23998fe