Increase DeepTDLearning act flexibility for state shape
Summary: This diff increases the flexibility of the `DeepTDLearning` policy learner for state shapes. The code change in `deep_td_learning.py` adds a check to handle cases where the subjective state has only one dimension, and then unsqueezes it to make it a batch of size 1. This allows the policy learner to handle state shapes of different dimensions.

Reviewed By: Yonathae

Differential Revision: D76654952

fbshipit-source-id: 89b2cdd48e91a0b6522bbf6786411438579316ee