Expand disjoint bandit test to have a subset of actions not observed in training data
Summary:
This diff expands the disjoint bandit test to include a subset of actions that are not observed in the training data.

The code changes include adding a new parameter `unobserved_actions_first_index` to specify the index of the first action that is not observed in the training data.

Because they are never observed in the training data, we expect that the _learned_ scores for these actions are identical (because the component linear bandits are all initialized zero weights, and never updated). This is checked by the test.

We set the ground truth of unobserved actions to match what the learned linear bandits will be (all weights being zero) so that the ground truth and learned models still agree about them, allowing us to testing that correct learning has taken place.

Reviewed By: frankchengmeta

Differential Revision: D76729861

fbshipit-source-id: 720164d9849b1aa54de8a2cb206b716fc5187b20