Implement per-row tie-breaking utility function
Summary: This diff implements a new utility function called `argmax_random_tie_break_per_row` which is used to break ties in per-row fashion. This function is used to randomly select an action from a set of ties in a row. The function is implemented by first finding the argmax of each row, and then randomly selecting an action from the ties in each row. The randomization is done independently for each row, unlike `argmax_random_tie_breaks` which uses the same permutation for all rows. This function is slower than `argmax_random_tiebreaks'

Reviewed By: frankchengmeta

Differential Revision: D76729863

fbshipit-source-id: f64a542c336618850f47b6b6aad3b8715eff53f2