Fix default terminated and truncated sizes in DataLoaderReplayBuffer
Summary:
DataLoaderReplayBuffer creates default values for `terminated` and `truncated` fields when columns for them are not provided.

We have recently identified a bug where there was a mismatch between these values' sizes and other values in the generated batch.

This was due to the fact that they were created with size equal to the `batch_size` parameter passed to method `sample`, whereas the batch itself was created according to the `batch_size` parameter passed to the `__init__` method of the class.

The bug has been revealed because unrelated recent changes caused these values to be different.

The solution is to use the same batch size value for all fields, which here means using the `batch_size` value passed to `__init__`. This is the value to be used because Korski dataframes require batch sizes to be set before iterating over their tables.

This diff adds some tests that catch this situation, and also improve the assertion error messages.

Reviewed By: PavlosApo

Differential Revision: D77912244

fbshipit-source-id: 02542507e60ef752abee33868a686004332d6ab1