Add `later` package dependency
Summary:
`later.unittest.TestCase` is now the recommended unit test class to use.

This corrects the absence of `later` in `pyproject.toml`.

It also changes an import from absolute to relative in `test_serialization.py` because that creates an error in the MacOS execution. See discussion in https://github.com/facebookresearch/Pearl/pull/116.

Differential Revision: D78047067